feature(embed): adds elgg/embed AMD module
Adds elgg/embed module that is defined as dependency of a menu item.
Deprecates elgg.embed JS library
